refactor: organize api helpers
